requirements/local.txt
# dependencies that are only related to deving on your local machine
-r test.txt

# Debugging
ipython==5.1.0

# Docs
mkdocs==0.15.3

# Command Utils
Fabric==1.12.0
pycrypto==2.6.1

coverage==4.4.1
django-coverage-plugin==1.5.0

Sphinx==1.6.3
django-extensions==1.8.1
Werkzeug==0.12.2
django-test-plus==1.0.18
factory-boy==2.9.2

django-debug-toolbar==1.8

# improved REPL
ipdb==0.10.3

pytest-django==3.1.2
pytest-sugar==0.9.0
